Core/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain_groq
httpx
pydantic
python-dotenv
requests
